requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
Django==3.1.3
djangorestframework==3.12.2
html2text==2020.1.16
idna==2.10
Pillow==8.0.1
pytz==2020.4
requests==2.25.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.26.2
